вынес методы в отдельный класс + поправил тест
