+++ spring-07/library/src/main/java/ru/otus/testing/service/impl/BookServiceImpl.java
@@ -2,15 +2,13 @@ package ru.otus.testing.service.impl;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.otus.testing.dao.AuthorRepository;
 import ru.otus.testing.dao.BookRepository;
-import ru.otus.testing.dao.CommentRepository;
-import ru.otus.testing.dao.GenreRepository;
 import ru.otus.testing.model.Author;
 import ru.otus.testing.model.Book;
 import ru.otus.testing.model.Comment;
 import ru.otus.testing.model.Genre;
 import ru.otus.testing.service.BookService;
+import ru.otus.testing.service.CheckDbFillingService;
 import ru.otus.testing.service.IOService;
 
 import java.util.List;
@@ -20,29 +18,23 @@ public class BookServiceImpl implements BookService {
 
     private final BookRepository bookRepository;
 
-    private final AuthorRepository authorRepository;
-
-    private final GenreRepository genreRepository;
-
-    private final CommentRepository commentRepository;
+    private final CheckDbFillingService checkDbFillingService;
 
     private final IOService ioService;
 
-    public BookServiceImpl(BookRepository bookRepository, AuthorRepository authorRepository,
-                           GenreRepository genreRepository, CommentRepository commentRepository, IOService ioService) {
+    public BookServiceImpl(BookRepository bookRepository, CheckDbFillingService checkDbFillingService,
+                           IOService ioService) {
         this.bookRepository = bookRepository;
-        this.authorRepository = authorRepository;
-        this.genreRepository = genreRepository;
-        this.commentRepository = commentRepository;
+        this.checkDbFillingService = checkDbFillingService;
         this.ioService = ioService;
     }
 
     @Transactional
     @Override
     public Book save(String bookName, long bookYear, Author author, Genre genre, List<Comment> commentsList) {
-        var authorInfoFromDb = findOrCreateAuthor(author);
-        var genreInfoFromDb = findOrCreateGenre(genre);
-        var commentsInfoFromDbList = findOrCreateComments(commentsList);
+        var authorInfoFromDb = checkDbFillingService.findOrCreateAuthor(author);
+        var genreInfoFromDb = checkDbFillingService.findOrCreateGenre(genre);
+        var commentsInfoFromDbList = checkDbFillingService.findOrCreateComments(commentsList);
 
         return bookRepository.save(new Book(bookName, bookYear, authorInfoFromDb, genreInfoFromDb,
                 commentsInfoFromDbList));
@@ -93,9 +85,9 @@ public class BookServiceImpl implements BookService {
         var findBook = bookRepository.findById(bookId);
 
         if (findBook.isPresent()) {
-            var authorInfoFromDb = findOrCreateAuthor(author);
-            var genreInfoFromDb = findOrCreateGenre(genre);
-            var commentsInfoFromDbList = findOrCreateComments(commentsList);
+            var authorInfoFromDb = checkDbFillingService.findOrCreateAuthor(author);
+            var genreInfoFromDb = checkDbFillingService.findOrCreateGenre(genre);
+            var commentsInfoFromDbList = checkDbFillingService.findOrCreateComments(commentsList);
 
             var presentFindBook = findBook.get();
             presentFindBook.setName(bookName);
@@ -111,52 +103,5 @@ public class BookServiceImpl implements BookService {
     @Override
     public void delete(long bookId) {
         bookRepository.deleteById(bookId);
-    }
-
-
-    /**
-     * Проверяет наличие автора в БД и при его отсутствии - создает
-     *
-     * @param author - автор книги
-     * @return возвращает модель автора - либо из БД, либо созданную и сохраненную
-     */
-    private Author findOrCreateAuthor(Author author) {
-        var authorInfoFromDb = authorRepository.findByNameAndYear(author.getName(), author.getYear());
-        if (authorInfoFromDb == null) {
-            authorInfoFromDb = authorRepository.save(author);
-        }
-        return authorInfoFromDb;
-    }
-
-    /**
-     * Проверяет наличие жанра в БД и при его отсутствии - создает
-     *
-     * @param genre - жанр книги
-     * @return возвращает модель жанра - либо из БД, либо созданный и сохраненный
-     */
-    private Genre findOrCreateGenre(Genre genre) {
-        var genreInfoFromDb = genreRepository.findByName(genre.getName());
-        if (genreInfoFromDb == null) {
-            genreInfoFromDb = genreRepository.save(genre);
-        }
-        return genreInfoFromDb;
-    }
-
-    /**
-     * Проверяет наличие комментариев в БД и при их отсутствии - создает
-     *
-     * @param commentsList - комментарии у книги
-     * @return возвращает лист комментариев - либо из БД, либо созданный и сохраненный
-     */
-    private List<Comment> findOrCreateComments(List<Comment> commentsList) {
-        var commentsInfoFromDbList = commentRepository.findByCommentsTextList(commentsList.stream()
-                .map(Comment::getCommentText).toList());
-        if (commentsInfoFromDbList.isEmpty()) {
-            for (var comment : commentsList) {
-                commentsInfoFromDbList.add(commentRepository.save(comment));
-            }
-        }
-
-        return commentsInfoFromDbList;
     }
 }
